backend/FastApi/requirements.txt
fastapi
uvicorn
numpy
opencv-python
Pillow
tensorflow
mediapipe
